refactor: move most of `dense_filter_util` and all of `group_by_util` to `base::database` - move most of `dense_filter_util` and all of `group_by_util` to `base::database` - rename the remaining file to `fold_util` - rename the moved file to `filter_util`
